Actually scan the image repository
This commit adds a little code to scan the image repository given in
an ImageRepository resource, and report the (number of) tags there. It
does so anonymously, so it'll only work for public repos for the
minute.

The next step is to introduce ImagePolicy resources and figure out the
interaction with them -- e.g., if they only care about semver, you
might only need to get the tags.