Be consistent about returning an error
It's mooted in https://github.com/fluxcd/toolkit/discussions/164 that
a distinct metric is used for not completing reconciliation, as
opposed to an unexpected error. Until that discussion has run its
course, we should just do what the other controllers do, and that's
returning an error when the controller is unable to reconcile to
completion.

This also adds a comment noting the purpose of the redundant `Requeue:
true` fields, for the avoidance of confusion later.