Respect the reconcileAt annotation
The GOTK controllers all use an annotation to trigger syncing outside
of the schedule (e.g., when a webhook is indicates a new
revision). This commit makes the ImageRepository controller respect
the annotation, by checking it when examining an image repo object.

This works a little different to the other controllers, which tend to
assume that any incoming event means a sync is due. The image
reflector controller uses the ChangedPredicate to filter incoming
evetns, as they do, but goes further to check the annotation and
otherwise, the scheduled scan time.