Fix ordering of copy and go mod in Dockerfile
Once the api directory has its own go.mod, which is refered to in the
top go.mod, it needs to be copied before running `go mod download` in
the Dockerfile.

This won't have been picked up in PR testing, because it's only in the
Dockerfile and the docker-build isn't run in the test workflow (maybe
it should be?).

Signed-off-by: Michael Bridgen <michael@weave.works>