Avoid race between DB reads and DB close in tests
Badger will panic with a nil pointer dereference if you do a
transaction after a database is closed. In the controller test suite,
the database is closed but the controllers, which read from it, are
left running.

To remove ths race, it is necessary to tell the controller manager
explicitly to shut down; this is done by cancelling the context given
to it.

Signed-off-by: Michael Bridgen <michael@weave.works>