Implement use of certs from a secret
This implements the design in #89, so you can provide a TLS
certificate and key to use for authentication; and, a CA certificate
for when you are using self-signed certs.

Unlike other GOTK controllers, the certificate secret is separate to
the credentials secret. This is because the credentials secret is more
likely to be a dockerconfigjson secret (i.e., created with

    docker create secret docker-registry ...

or its equivalent).

Signed-off-by: Michael Bridgen <michael@weave.works>