Add registry login integration test for AWS EKS
Use tftestenv to create tests similar to the controller tests but with
EKS cluster provisioned using terraform. The test setup is written
similar to the usual controller test suite_test.go, providing
flexibility in the test setup and execution. The structure tries to keep
the infrastructure code separate from the test code such that the
infrastructure provisioning tool can be easily replaced, without the
need to change the actual tests.

The go tests are also written similar to the usual controller tests,
with a kube client from the tftestenv environment global variable.

Terraform is automatically downloaded using the terraform-exec package.
The test container images are generated using the go-containerregistry
package and pushed to the test container image repository.

Co-authored-by: Somtochi Onyekwere <somtochionyekwere@gmail.com>
Signed-off-by: Sunny <darkowlzz@protonmail.com>