tests/integration: Add GCP GKE tests
Add GCP as a new provider to provision infrastructure and test against.

GCP has Google Container Registry and Artifact Registry for storing
images. The test setup provisions and pushes images to both and tests
them both.

The GKE node pool is configured with OAuth Scope to provide full access
to the GCP APIs. Workload identity is not used to avoid increasing the
complixity of the setup.

Co-authored-by: Somtochi Onyekwere <somtochionyekwere@gmail.com>
Signed-off-by: Sunny <darkowlzz@protonmail.com>