cloud-provider-e2e: Use test image-reflector build
Update the test suite to take the image-reflector-controller build to be
tested using the variable TEST_IMG. This test image must be built
locally before running the e2e tests. In the e2e tests, this local image
is retagged and pushed into a remote registry and is used in flux
deployment.

For Google Artifact Registry and Azure Container Registry, the
previously created test registries are reused to push the
image-reflector image. But for ECR, a new repository is created due to
lack of support for dynamic repositories.

Signed-off-by: Sunny <darkowlzz@protonmail.com>