Refactor ImageRepository and ImagePolicy
Refactor the two reconcilers to add support for kstatus based status
conditions and break down the reconciler into smaller easily testable
units.
It also fixes a lot of hidden bugs in the behavior of the reconcilers.

A few of the fixed issues and change in behavior are:
- Avoid writing to the ImageRepository spec for adding default exclusion
  list values. The default is now set via the CRD API.
- Bug in the exclusion list filter. Multiple patterns in the list didn't
  filter the tags for all the patterns.
- Updating image or the exclusion list of ImageRepository didn't force
  scan. It waited for the scan period before scanning. Now, if they are
  updated, scan is performed immediately.
- For an existing ImageRepository, if the image is changed, the old
  scan result is now deleted as that no longer belongs to the new
  image.

Also, silence badger and test registry server logs.

Signed-off-by: Sunny <darkowlzz@protonmail.com>