add label sharding capability to controller
The new `--watch-label-selector` lets operators provide a label to the
controller manager which in turn uses it to reconcile only those
resources (ImageRepositories and ImagePolicies) that match the given
label.

This way operators can deploy multiple instances of IRC, each
reconciling a distinct set of resources based on their labels and
effectively scale IRC horizontally.

closes #364

Signed-off-by: Max Jonas Werner <mail@makk.es>